Fix integer overflow bug in DataTypes

+++ src/DataTypes.java
@@ -3,7 +3,7 @@ public class DataTypes {
     public static long sum(int[] numbers) {
         long s = 0; // variable to accumulate the sum in!// changed to long to avoid integer overflow.
         // below is a "foreach" loop which iterates through numbers
-        for (long x : numbers) {
+        for (int x : numbers) {
             s += x;
         }
         return s;
